+++ src/App/Documentation/components/Dialog/index.js
@@ -33,7 +33,7 @@ const Example = () => (
     <>
         <h2 id="overview">Example</h2>
         <ComponentPreview language="html" showCasePanel codeFigure>
-            <button className="btn btn-executive" type="button" data-dialog-open="demo-dialog">Open dialog</button>
+            <button className="btn btn-executive" type="button" data-dialog-open="demo-dialog">{"\n"}Open dialog{"\n"}</button>
             <DialogComponent diaHeader="Delete item 456?" diaId="demo-dialog">
                 <p>Are you sure you want to permanently delete the item <i>German Swashbuckle (456)?</i></p>
             </DialogComponent>