+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Added .dialog-header and .dialog-footer to the dialog component.
 - Added tab lock to dialog making it so you can't tab out of the dialog until an option is clicked.
+- Added aria labels to dialog documentation to show an example of how to use dialog accessibly.
 
 ### Changed